chore(pages): move cards component into a single file
